pyproject.toml
[project]
name = "django-boilerplate"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "434mdhuzaifa", email = "huzaifa@workspaceit.com"},
]
dependencies = [
    "django==5",
    "djangorestframework>=3.15.2",
    "django-cors-middleware>=1.5.0",
    "icecream>=2.1.3",
    "python-dotenv>=1.0.1",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.scripts]
start.cmd="manage.py runserver"
stat.env_file =".env"
makemigrations="manage.py makemigrations"
migrate="manage.py migrate"
startapp="django-admin startapp"